changed the package name of accountmetadata to prevent having service 3 times
