+++ components/account-metadata-service/pom.xml
@@ -25,8 +25,8 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <artifactId>reference-implementation-consumerdatastandards-au</artifactId>
     <groupId>org.wso2.openbanking</groupId>
+    <artifactId>reference-implementation-consumerdatastandards-au</artifactId>
     <version>2.0.0</version>
     <relativePath>../../pom.xml</relativePath>
   </parent>
@@ -35,50 +35,21 @@
   <name>WSO2 Open Banking Toolkit - Account Metadata API</name>
   <packaging>war</packaging>
 
-  <properties>
-    <jackson.version>2.17.1</jackson.version>
-
-    <orgjson.version>20240303</orgjson.version>
-
-    <mockito.version>5.11.0</mockito.version>
-    <testng.version>7.10.2</testng.version>
-  </properties>
-
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>com.fasterxml.jackson</groupId>
-        <artifactId>jackson-bom</artifactId>
-        <version>${jackson.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-
-      <dependency>
-        <groupId>org.json</groupId>
-        <artifactId>json</artifactId>
-        <version>${orgjson.version}</version>
+        <groupId>commons-codec</groupId>
+        <artifactId>commons-codec</artifactId>
+        <version>${commons.codec.version}</version>
       </dependency>
     </dependencies>
   </dependencyManagement>
 
   <dependencies>
+
     <dependency>
       <groupId>commons-logging</groupId>
       <artifactId>commons-logging</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>com.fasterxml.jackson.core</groupId>
-      <artifactId>jackson-databind</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.datatype</groupId>
-      <artifactId>jackson-datatype-jsr310</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.jaxrs</groupId>
-      <artifactId>jackson-jaxrs-json-provider</artifactId>
     </dependency>
 
     <dependency>
@@ -86,48 +57,49 @@
       <artifactId>javax.servlet-api</artifactId>
       <scope>provided</scope>
     </dependency>
+
     <dependency>
-      <groupId>jakarta.ws.rs</groupId>
-      <artifactId>jakarta.ws.rs-api</artifactId>
+      <groupId>javax.ws.rs</groupId>
+      <artifactId>javax.ws.rs-api</artifactId>
       <scope>provided</scope>
     </dependency>
 
     <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-lang3</artifactId>
+      <version>${commons.lang3.version}</version>
     </dependency>
+
     <dependency>
       <groupId>org.apache.httpcomponents</groupId>
       <artifactId>httpclient</artifactId>
-      <version>4.5.14</version>
+      <version>${httpclient.version}</version>
     </dependency>
 
     <dependency>
       <groupId>org.json</groupId>
       <artifactId>json</artifactId>
+      <version>${org.json.version}</version>
     </dependency>
 
     <dependency>
       <groupId>com.google.code.gson</groupId>
       <artifactId>gson</artifactId>
     </dependency>
+
     <dependency>
       <groupId>io.swagger</groupId>
       <artifactId>swagger-annotations</artifactId>
     </dependency>
+
     <dependency>
       <groupId>javax.validation</groupId>
       <artifactId>validation-api</artifactId>
     </dependency>
+
     <dependency>
       <groupId>javax.annotation</groupId>
       <artifactId>javax.annotation-api</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.ops4j.pax.logging</groupId>
-      <artifactId>pax-logging-api</artifactId>
-      <version>2.1.0</version>
-      <scope>compile</scope>
     </dependency>
 
     <dependency>
@@ -139,17 +111,23 @@
 
     <dependency>
       <groupId>org.mockito</groupId>
-      <artifactId>mockito-core</artifactId>
-      <version>${mockito.version}</version>
+      <artifactId>mockito-all</artifactId>
       <scope>test</scope>
     </dependency>
 
     <dependency>
       <groupId>org.glassfish.jersey.core</groupId>
       <artifactId>jersey-common</artifactId>
-      <version>2.35</version>
+      <version>${jersey.common.version}</version>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>com.fasterxml.jackson.core</groupId>
+      <artifactId>jackson-annotations</artifactId>
+      <version>2.19.2</version>
+      <scope>compile</scope>
+    </dependency>
+
   </dependencies>
 
   <build>
@@ -157,7 +135,7 @@
       <plugin>
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>build-helper-maven-plugin</artifactId>
-        <version>3.3.0</version>
+        <version>${build.helper.maven.plugin.version}</version>
         <executions>
           <execution>
             <phase>generate-sources</phase>
@@ -187,21 +165,9 @@
       </plugin>
 
       <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.11.0</version>
-        <configuration>
-          <source>11</source>
-          <target>11</target>
-        </configuration>
-      </plugin>
-
-      <plugin>
         <groupId>org.jacoco</groupId>
         <artifactId>jacoco-maven-plugin</artifactId>
-        <version>0.8.11</version>
         <executions>
-
           <execution>
             <id>prepare-agent</id>
             <goals>
@@ -251,5 +217,16 @@
       </plugin>
     </plugins>
   </build>
+
+  <properties>
+    <build.helper.maven.plugin.version>3.3.0</build.helper.maven.plugin.version>
+    <jersey.common.version>2.35</jersey.common.version>
+    <pax.logging.api.version>2.1.0</pax.logging.api.version>
+    <commons.lang3.version>3.18.0</commons.lang3.version>
+    <httpclient.version>4.5.14</httpclient.version>
+    <org.json.version>20231013</org.json.version>
+    <testng.version>7.7.0</testng.version>
+    <commons.codec.version>1.21.0</commons.codec.version>
+  </properties>
 
 </project>
