+++ components/account-metadata-service/src/gen/java/org/wso2/openbanking/consumerdatastandards/account/metadata/model/DisclosureOptionItem.java
@@ -1,9 +1,8 @@
-package org.wso2.openbanking.consumerdatastandards.account.metadata.service.model;
+package org.wso2.openbanking.consumerdatastandards.account.metadata.model;
 
 import io.swagger.annotations.ApiModelProperty;
 import javax.validation.constraints.*;
 
-import io.swagger.annotations.*;
 import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
