+++ components/account-metadata-service/src/main/java/org/wso2/openbanking/consumerdatastandards/account/metadata/configurations/ConfigurableProperties.java
@@ -16,15 +16,13 @@
  * under the License.
  */
 
-package org.wso2.openbanking.consumerdatastandards.account.metadata.service.constants;
+package org.wso2.openbanking.consumerdatastandards.account.metadata.configurations;
 
 /**
- * This class holds common constants for the CDS Open Banking implementation.
+ * This class holds configurable properties for the CDS Open Banking implementation.
  */
-public class CommonConstants {
+public class ConfigurableProperties {
 
-    //Constants Related to Joint Accounts
-    public static final String DOMS_STATUS_PRE_APPROVAL = "pre-approval";
-    public static final String DOMS_STATUS_NO_SHARING = "no-sharing";
+    public static final String ACCOUNT_METADATA_DATASOURCE_JNDI_NAME = "jdbc/ACCOUNT_METADATA_DB";
 
 }
