+++ components/account-metadata-service/src/main/java/org/wso2/openbanking/consumerdatastandards/account/metadata/utils/connection/provider/ConnectionProvider.java
@@ -16,13 +16,13 @@
  * under the License.
  */
 
-package org.wso2.openbanking.consumerdatastandards.account.metadata.service.utils.connection.provider;
+package org.wso2.openbanking.consumerdatastandards.account.metadata.utils.connection.provider;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 
 /**
- * Provides database connections to the service layer.
+ * Provides database connections to the core layer.
  * This abstraction allows the underlying connection retrieval mechanism
  * (e.g., direct JDBC connection, connection pooling, JNDI datasource)
  * to be changed without impacting the business logic. It also improves
