+++ components/demo-backend/src/main/java/org/wso2/openbanking/cds/demo/backend/services/BankingService.java
@@ -639,7 +639,7 @@ public class BankingService {
                 "        \"endToEndId\": \"0987\",\n" +
                 "        \"purposeCode\": \"0456\"\n" +
                 "      },\n" +
-                "      \"service\": \"X2P1.01\"\n" +
+                "      \"core\": \"X2P1.01\"\n" +
                 "    }\n" +
                 "  }";
         return jsonParser.parse(transactionJson).getAsJsonObject();
