+++ components/reference-implementation-openbanking-consumerdatastandards-au/src/main/java/org/wso2/openbanking/consumerdatastandards/au/extensions/gen/model/RequestForPreProcessFileUpload.java
@@ -106,7 +106,7 @@ public class RequestForPreProcessFileUpload   {
   }
 
   /**
-   * Request headers sent by the TPP. Filtered set of headers are sent to the external service.
+   * Request headers sent by the TPP. Filtered set of headers are sent to the external core.
    **/
   public RequestForPreProcessFileUpload requestHeaders(Object requestHeaders) {
     this.requestHeaders = requestHeaders;
@@ -114,7 +114,7 @@ public class RequestForPreProcessFileUpload   {
   }
 
   
-  @ApiModelProperty(value = "Request headers sent by the TPP. Filtered set of headers are sent to the external service.")
+  @ApiModelProperty(value = "Request headers sent by the TPP. Filtered set of headers are sent to the external core.")
   @JsonProperty("requestHeaders")
   public Object getRequestHeaders() {
     return requestHeaders;
