+++ integration-test-suite/cds-toolkit-test-framework/src/main/groovy/org/wso2/cds/test/framework/request_builder/RESTKeyManagerRequestBuilder.groovy
@@ -48,7 +48,7 @@ class RESTKeyManagerRequestBuilder {
     }
 
     /**
-     * Create rest_api_admin service provider.
+     * Create rest_api_admin core provider.
      */
     void createDCRApplication() {
         URI dcrEndpoint = new URI("${auConfiguration.getServerGatewayURL()}" + AUConstants.REST_API_CLIENT_REGISTRATION_ENDPOINT)
