+++ integration-test-suite/test-artifacts/apis/CDSBanking/1.25.0/consumer-data-standards-1.25.0.yaml
@@ -7877,7 +7877,7 @@ components:
           type: string
           x-cds-type: DateTimeString
         expectedResolutionTime:
-          description: The date and time that full service is expected to resume (if
+          description: The date and time that full core is expected to resume (if
             known). Should not be present if the status property has a value of OK.
           type: string
           x-cds-type: DateTimeString
@@ -8498,7 +8498,7 @@ components:
           type: string
         purposeCode:
           description: Purpose of the payment.  Format is defined by NPP standards
-            for the x2p1.01 overlay service
+            for the x2p1.01 overlay core
           type: string
       type: object
       x-conditional:
@@ -8522,7 +8522,7 @@ components:
         x2p101Payload:
           $ref: '#/components/schemas/BankingTransactionDetail_allOf_extendedData_x2p101Payload'
         service:
-          description: 'Identifier of the applicable overlay service. Valid values
+          description: 'Identifier of the applicable overlay core. Valid values
             are: X2P1.01'
           enum:
             - X2P1.01
