+++ toolkits/ob-apim/repository/resources/apis/consumerdatastandards.org.au/CDSBanking/1.8.0/consumer-data-standards-1.8.0.yaml
@@ -3790,7 +3790,7 @@ definitions:
         description: An end to end ID for the payment created at initiation
       purposeCode:
         type: string
-        description: Purpose of the payment.  Format is defined by NPP standards for the x2p1.01 overlay service
+        description: Purpose of the payment.  Format is defined by NPP standards for the x2p1.01 overlay core
   BankingTransactionDetail_extendedData:
     required:
       - service
@@ -3810,7 +3810,7 @@ definitions:
         $ref: '#/definitions/BankingTransactionDetail_extendedData_x2p101Payload'
       service:
         type: string
-        description: 'Identifier of the applicable overlay service. Valid values are: X2P1.01'
+        description: 'Identifier of the applicable overlay core. Valid values are: X2P1.01'
         enum:
           - X2P1.01
   ResponseBankingAccountsBalanceList_data:
@@ -3930,7 +3930,7 @@ definitions:
         x-cds-type: DateTimeString
       expectedResolutionTime:
         type: string
-        description: The date and time that full service is expected to resume (if known). Should not be present if the status property has a value of OK.
+        description: The date and time that full core is expected to resume (if known). Should not be present if the status property has a value of OK.
         x-cds-type: DateTimeString
       updateTime:
         type: string
